Restructure client and session
