+++ odoo/client_test.go
@@ -3,148 +3,142 @@ package odoo
 import (
 	"context"
 	"io/ioutil"
+	"math/rand"
 	"net/http"
 	"net/http/httptest"
+	"strconv"
 	"testing"
 
 	"github.com/go-logr/logr"
 	"github.com/go-logr/zapr"
+	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap/zapcore"
 	"go.uber.org/zap/zaptest"
 )
 
-func TestClient_CreateGenericModel(t *testing.T) {
-	numRequests := 0
-	uuidGenerator = func() string {
-		return "fakeID"
-	}
+func TestClient_Login_Success(t *testing.T) {
+	var (
+		numRequests  = 0
+		testLogin    = uuid.NewString()
+		testPassword = uuid.NewString()
+		testUID      = rand.Int()
+		testSID      = uuid.NewString()
+	)
+
 	odooMock := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		numRequests++
-		assert.Equal(t, "/web/dataset/call_kw/create", r.RequestURI)
+		assert.Equal(t, "/web/session/authenticate", r.RequestURI)
 
-		buf, err := ioutil.ReadAll(r.Body)
-		assert.NoError(t, err)
-		assert.JSONEq(t, `{
-			"id":"fakeID",
-			"jsonrpc":"2.0",
-			"method":"call",
-			"params":{
-				"model":"model",
-				"method":"create",
-				"args":[					
-					"data"
-				],
-				"kwargs":{}
-			}}`, string(buf))
+		b, err := ioutil.ReadAll(r.Body)
+		require.NoError(t, err)
+		body := string(b)
+		assert.Contains(t, body, `"db":"TestDB"`)
+		assert.Contains(t, body, `"login":"`+testLogin+`"`)
+		assert.Contains(t, body, `"password":"`+testPassword+`"`)
 
 		w.Header().Set("content-type", "application/json")
 		_, err = w.Write([]byte(`{
+			"id": "1337",
 			"jsonrpc": "2.0",
-			"id": "fakeID",
-			"result": 221
+			"result": {
+				"company_id": 1,
+				"db": "TestDB",
+				"session_id": "` + testSID + `",
+				"uid": ` + strconv.Itoa(testUID) + `,
+				"user_context": {
+					"lang": "en_US",
+					"tz": "Europe/Zurich",
+					"uid": ` + strconv.Itoa(testUID) + `
+				},
+				"username": "` + testLogin + `"
+			}
 		}`))
 		require.NoError(t, err)
 	}))
 	defer odooMock.Close()
 
-	// Do request
-	client := NewClient(odooMock.URL, "TestDB")
-	client.UseDebugLogger(true)
-	m := NewCreateModel("model", "data")
-	result, err := client.CreateGenericModel(newTestContext(t), &Session{}, m)
+	// Login
+	client, err := NewClient(odooMock.URL, "TestDB")
 	require.NoError(t, err)
-	assert.Equal(t, 221, result)
+	client.UseDebugLogger(true)
+	session, err := client.Login(newTestContext(t), testLogin, testPassword)
+	require.NoError(t, err)
+	assert.Equal(t, testUID, session.UID)
+	assert.Equal(t, testSID, session.SessionID)
 	assert.Equal(t, 1, numRequests)
 }
 
-func TestClient_UpdateGenericModel(t *testing.T) {
-	numRequests := 0
-	uuidGenerator = func() string {
-		return "fakeID"
-	}
+func TestLogin_BadCredentials(t *testing.T) {
+	var (
+		numRequests  = 0
+		testLogin    = uuid.NewString()
+		testPassword = uuid.NewString()
+		testSID      = uuid.NewString()
+	)
+
 	odooMock := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		numRequests++
-		assert.Equal(t, "/web/dataset/call_kw/write", r.RequestURI)
-
-		buf, err := ioutil.ReadAll(r.Body)
-		assert.NoError(t, err)
-		assert.JSONEq(t, `{
-			"id":"fakeID",
-			"jsonrpc":"2.0",
-			"method":"call",
-			"params":{
-				"model":"model",
-				"method":"write",
-				"args":[
-					[1],
-					"data"
-				],
-				"kwargs":{}
-			}}`, string(buf))
-
+		assert.Equal(t, "/web/session/authenticate", r.RequestURI)
 		w.Header().Set("content-type", "application/json")
-		_, err = w.Write([]byte(`{
+		_, err := w.Write([]byte(`{
+			"id": "1337",
 			"jsonrpc": "2.0",
-			"id": "fakeID",
-			"result": true
+			"result": {
+				"company_id": null,
+				"db": "TestDB",
+				"session_id": "` + testSID + `",
+				"uid": false,
+				"user_context": {},
+				"username": "` + testLogin + `"
+			}
 		}`))
 		require.NoError(t, err)
 	}))
 	defer odooMock.Close()
 
 	// Do request
-	client := NewClient(odooMock.URL, "TestDB")
+	client, err := NewClient(odooMock.URL, "TestDB")
+	require.NoError(t, err)
 	client.UseDebugLogger(true)
-	m, err := NewUpdateModel("model", 1, "data")
-	require.NoError(t, err)
-	result, err := client.UpdateGenericModel(newTestContext(t), &Session{}, m)
-	require.NoError(t, err)
-	assert.True(t, result)
+	session, err := client.Login(newTestContext(t), testLogin, testPassword)
+	require.EqualError(t, err, "invalid credentials")
+	assert.Nil(t, session)
 	assert.Equal(t, 1, numRequests)
 }
 
-func TestClient_DeleteGenericModel(t *testing.T) {
+func TestLogin_BadResponse(t *testing.T) {
 	numRequests := 0
-	uuidGenerator = func() string {
-		return "fakeID"
-	}
 	odooMock := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		numRequests++
-		assert.Equal(t, "/web/dataset/call_kw/unlink", r.RequestURI)
-
-		buf, err := ioutil.ReadAll(r.Body)
-		assert.NoError(t, err)
-		assert.JSONEq(t, `{
-			"id":"fakeID",
-			"jsonrpc":"2.0",
-			"method":"call",
-			"params":{
-				"model":"model",
-				"method":"unlink",
-				"args":[[100]],
-				"kwargs":{}
-			}}`, string(buf))
 
 		w.Header().Set("content-type", "application/json")
-		_, err = w.Write([]byte(`{
+		_, err := w.Write([]byte(`{
 			"jsonrpc": "2.0",
-			"id": "fakeID",
-			"result": true
-		}`))
+			"id": "xxx",
+			"error": {
+			  "message": "Odoo Server Error",
+			  "code": 200,
+			  "data": {
+				"debug": "Traceback xxx",
+				"message": "",
+				"name": "werkzeug.exceptions.Foo",
+				"arguments": []
+			  }
+			}
+		  }`))
 		require.NoError(t, err)
 	}))
 	defer odooMock.Close()
 
 	// Do request
-	client := NewClient(odooMock.URL, "TestDB")
+	client, err := NewClient(odooMock.URL, "TestDB")
+	require.NoError(t, err)
 	client.UseDebugLogger(true)
-	m, err := NewDeleteModel("model", []int{100})
-	require.NoError(t, err)
-	result, err := client.DeleteGenericModel(newTestContext(t), &Session{}, m)
-	require.NoError(t, err)
-	assert.True(t, result)
+	session, err := client.Login(newTestContext(t), "", "")
+	require.EqualError(t, err, "error from Odoo: &{Odoo Server Error 200 map[arguments:[] debug:Traceback xxx message: name:werkzeug.exceptions.Foo]}")
+	assert.Nil(t, session)
 	assert.Equal(t, 1, numRequests)
 }
