+++ odoo/model/odoo.go
@@ -2,15 +2,14 @@ package model
 
 import "github.com/vshn/appuio-odoo-adapter/odoo"
 
+// Odoo is the developer-friendly odoo.Client with strongly-typed models.
 type Odoo struct {
-	client  *odoo.Client
 	session *odoo.Session
 }
 
 // NewOdoo creates a new Odoo client.
-func NewOdoo(session *odoo.Session, client *odoo.Client) *Odoo {
+func NewOdoo(session *odoo.Session) *Odoo {
 	return &Odoo{
-		client:  client,
 		session: session,
 	}
 }
