+++ odoo/session.go
@@ -2,9 +2,9 @@ package odoo
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
+	"io"
 	"net/http"
 )
 
@@ -19,70 +19,82 @@ type Session struct {
 	// Is always set, no matter the authentication outcome.
 	SessionID string `json:"session_id,omitempty"`
 	// UID is the user's ID as an int, or the boolean `false` if authentication failed.
-	UID int `json:"uid,omitempty"`
+	UID    int `json:"uid,omitempty"`
+	client *Client
 }
 
-type loginParams struct {
-	DB       string `json:"db,omitempty"`
-	Login    string `json:"login,omitempty"`
-	Password string `json:"password,omitempty"`
+// SearchGenericModel accepts a SearchReadModel and unmarshal the response into the given pointer.
+// Depending on the JSON fields returned a custom json.Unmarshaler needs to be written since Odoo sets undefined fields to `false` instead of null.
+func (c *Session) SearchGenericModel(ctx context.Context, model SearchReadModel, into interface{}) error {
+	return c.executeGenericRequest(ctx, c.client.parsedURL.String()+"/web/dataset/search_read", model, into)
 }
 
-// Login tries to authenticate the user against Odoo.
-// It returns a session if authentication was successful. An error is returned if
-//  - the credentials were wrong,
-//  - encoding or sending the request,
-//  - or decoding the request failed.
-func (c Client) Login(ctx context.Context, login, password string) (*Session, error) {
-	resp, err := c.requestSession(ctx, login, password)
-	if err != nil {
-		return nil, err
+// CreateGenericModel accepts a WriteModel as a payload and executes a query to create the new data record.
+func (c *Session) CreateGenericModel(ctx context.Context, model WriteModel) (int, error) {
+	if model.KWArgs == nil {
+		model.KWArgs = map[string]interface{}{} // set to non-null when serializing
 	}
-
-	return c.decodeSession(resp)
+	resultID := 0
+	err := c.executeGenericRequest(ctx, c.client.parsedURL.String()+"/web/dataset/call_kw/create", model, &resultID)
+	return resultID, err
 }
 
-func (c Client) requestSession(ctx context.Context, login string, password string) (*http.Response, error) {
-	// Prepare request
-	body, err := NewJSONRPCRequest(loginParams{c.db, login, password}).Encode()
-	if err != nil {
-		return nil, newEncodingRequestError(err)
+// UpdateGenericModel accepts a WriteModel as a payload and executes a query to update an existing data record.
+func (c *Session) UpdateGenericModel(ctx context.Context, model WriteModel) (bool, error) {
+	if model.KWArgs == nil {
+		model.KWArgs = map[string]interface{}{} // set to non-null when serializing
 	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, c.parsedURL.String()+"/web/session/authenticate", body)
-	if err != nil {
-		return nil, newCreatingRequestError(err)
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	// Send request
-	resp, err := c.http.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("login: sending HTTP request: %w", err)
-	}
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("login: expected HTTP status 200 OK, got %s", resp.Status)
-	}
-	return resp, nil
+	updated := false
+	err := c.executeGenericRequest(ctx, c.client.parsedURL.String()+"/web/dataset/call_kw/write", model, &updated)
+	return updated, err
 }
 
-func (c Client) decodeSession(res *http.Response) (*Session, error) {
-	// Decode response
-	// We don't use DecodeResult here because we're interested in whether unmarshalling the result failed.
-	// If so, this is likely because "uid" is set to `false` which indicates an authentication failure.
-	var response JSONRPCResponse
-	if err := json.NewDecoder(res.Body).Decode(&response); err != nil {
-		return nil, fmt.Errorf("login: decode response: %w", err)
+// DeleteGenericModel accepts a WriteModel as a payload and executes a query to delete an existing data record.
+// For the query to succeed it is required that the Model sets an ID.
+func (c *Session) DeleteGenericModel(ctx context.Context, model WriteModel) (bool, error) {
+	if model.KWArgs == nil {
+		model.KWArgs = map[string]interface{}{} // set to non-null when serializing
 	}
-	if response.Error != nil {
-		return nil, fmt.Errorf("error from Odoo: %v", response.Error)
+	deleted := false
+	err := c.executeGenericRequest(ctx, c.client.parsedURL.String()+"/web/dataset/call_kw/unlink", model, &deleted)
+	return deleted, err
+}
+
+func (c *Session) executeGenericRequest(ctx context.Context, url string, model interface{}, into interface{}) error {
+	body, err := NewJSONRPCRequest(&model).Encode()
+	if err != nil {
+		return newEncodingRequestError(err)
 	}
 
-	// Decode session
-	var session Session
-	if err := json.Unmarshal(*response.Result, &session); err != nil {
-		// UID is not set, authentication failed
-		return nil, ErrInvalidCredentials
+	// Create request
+	req, err := http.NewRequestWithContext(ctx, http.MethodPost, url, body)
+	if err != nil {
+		return newCreatingRequestError(err)
 	}
+	req.Header.Set("content-type", "application/json")
+	req.Header.Set("cookie", "session_id="+c.SessionID)
 
-	return &session, nil
+	resp, err := c.sendRequest(req)
+	if err != nil {
+		return err
+	}
+	return c.unmarshalResponse(resp.Body, into)
+}
+
+func (c *Session) sendRequest(req *http.Request) (*http.Response, error) {
+	res, err := c.client.http.Do(req)
+	if err != nil {
+		return nil, fmt.Errorf("sending HTTP request: %w", err)
+	} else if res.StatusCode != http.StatusOK {
+		return nil, fmt.Errorf("expected HTTP status 200 OK, got %s", res.Status)
+	}
+	return res, nil
+}
+
+func (c *Session) unmarshalResponse(body io.ReadCloser, into interface{}) error {
+	defer body.Close()
+	if err := DecodeResult(body, into); err != nil {
+		return fmt.Errorf("decoding result: %w", err)
+	}
+	return nil
 }
